+++ determinant.h
@@ -9,7 +9,7 @@ class Determinant
 public:
     void Swap(int &a, int &b);
     void Perm(QVector<int> vec, QVector<QVector<int> > *vec_seq, int begin, int end);
-    float calculate(Eigen::Matrix<float, 2, 2> M);
+    double calculate(Eigen::Matrix<double, 2, 2> M);
     bool Iseven(int num);
     bool PowerIsPosition(QVector<int> & vec);
     QVector<int> inivec(int n);
